BCMVideo: Disable gamma on Pi 4
The Raspberry Pi 4 implements gamma support in a different method to
earlier models. Attempting to use the pre-existing gamma mailbox message
on Pi 4 will result in incorrect gamma settings being used (black stays
black, but everything else will become lighter & washed-out). It's
unclear when the Pi foundation will fix the firmware and/or document the
new gamma method, so for now our best option is to disable gamma on Pi
4.

Ref: https://github.com/raspberrypi/firmware/issues/1274

Tested on Pi 3 & 4