ToolboxLib: Const correctness of arg to toolbox_template_lookup()
The pointer whose address is passed in argument `hdr` will, on exit, point into
memory that is managed by the Toolbox module, and which therefore should not be
written by the callee (for example the module should be within its rights to
re-use the same location to service another similar call). The pointer should
therefore have type `const ObjectTemplateHeader *`, so `hdr` should have type
`const ObjectTemplateHeader **` to save well-behaved clients from having to
use a typecast.

Version 1.18. Tagged as 'ToolboxLib-1_18'