HAL_BCM2835: 8GB RAM support
Update OS_AddRAM sequence to detect 8GB Pi models and register the extra
RAM bank.

Also, remove the software RAM clear code; it won't work for clearing RAM
above the 4GB barrier, and (for a software RAM clear) it's quicker for
the kernel to clear the RAM itself (since it'll do it with the
MMU/caches enabled, and will do it in parallel with the keyboard scan).
E.g. for a 4GB Pi, the time between the HalStartup debug message (prior
to HAL RAM clear) and ModuleInit debug message (after kernel RAM clear)
used to be 1124cs, but is now 636cs. For an 8GB Pi the time is 948cs.

Version 0.94. Tagged as 'HAL_BCM2835-0_94'