HAL_BCM2835: Don't return Pi 4-only GPIO in all enumerations
The pin list table had the Pi 4-only alternates mixed in, the result of which is that enumerating the pins on something other than Pi 4 would return ghost alternates for peripherals that don't exist.
Split the pin lists and pick the appropriate one.

Also, add 2 missing board revisions that have sneaked out for 2B/3B.