SCSIFS: [464] Restore documented enumeration SWI behaviour
SCSI_Partitions 1 was changed in SCSIFS 1.34 to tie new significance to R2, which was previously undefined. Thus any clients of this subreason failed in various amusing ways (the ticket references SystemDisc and Fat32Formatte, and inspection of HForm suggests this would have altered behaviour because BASIC zeros unused registers).

Since the only client of this extension is the SCSI specific PartMan, and the module already has a variety of assembly time switches in order to complete the other half of the dialogue to work out which partition is which, just make the significant R2 depend on that switch too. Therefore the previously documented SWI behaviour is restored otherwise.

Also, tidy the docs up, and paste those tidied docs into the relevant place in the sources.

Version 1.36. Tagged as 'SCSIFS-1_36'