RISC_OSLib: Enable modulewrapped applications to use ResGen
This utilises a weak symbol import so that if the symbol `Resources` resolves
at link time, it is called to discover the address of the resource file data
block(s) to be installed in ResourceFS. If this fails, it drops back to using
the `s.ResFiles` source file as previously, thereby ensuring compatibility
with components that have not yet been converted to use ResGen.

Since `Resources` is the default area name (and thus discovery function name)
used for ResGen in the shared makefiles, this makes it as simple as possible
to utilise it.

Version 6.08. Tagged as 'RISC_OSLib-6_08'