HForm: Find occupied drives after empty ones and warn if silly
Change the loop scanning for SCSI discs to not stop if a LUN reports an error, only stop if the LUN is not present. This ensures that multi-in-one card readers that have empty sockets in the lower numbered LUNs (giving "Target error - not ready" errors) don't cause the discs in higher numbered LUNs to be missed.

Also, based on an idea from Alan Adams, warn if it appears the user is about to do something silly. That is: formatting the !Boot disc or formatting the disc on which !HForm is running. Neither is illegal, but it may reduce the chance of someone who didn't read the previous 9 prompts saying what's about to happen to think harder before the 10th and final prompt.

Version 2.76. Tagged as 'HForm-2_76'