SaveAs: Don't close save box when raising "To save drag..."
Detail:
The behaviour of Draw, Edit and Paint is for the save box to remain
open after reporting "To save, drag...", this commit changes SaveAs
to do the same.

Changed Germany and UK Res files so that all "Save" buttons
have local flag ticked, otherwise the box closes regardless of
wheather SaveInternal_CloseWhenFinished is used.
In events.c, event_action_button_save, when a filename without
path has been detected, clear the CloseWhenFinished flag, so that
the box remains open.

Admin:
The correct behaviour was observed in a Res file loaded with
ResTest.

Version 0.21. Tagged as 'SaveAs-0_21'