StringLib: Channel 8 RMA corruption fix
Adjust the workspace definitions so that Instance0 has the same
address/offset as End_Of_Code, and so that WorkSpaceSize has the value
Instance0 + 8*4.

The previous version of the code made the false assumption that the
AmountToCopy and WorkSpaceSize definitions wouldn't add anything to the
output binary. This resulted in WorkSpaceSize being 4 bytes larger than
strictly necessary (due to AmountToCopy before it), but more
dangerously, Instance0 was placed 8 bytes after End_Of_Code, causing the
code to write off the end of the heap block if the voice is assigned to
channel 8.

Fixes https://www.riscosopen.org/tracker/tickets/505 (StringLib variant)

Version 1.20. Tagged as 'StringLib-1_20'